+++ cpp/parser.hpp
@@ -1,6 +1,6 @@
 //
 //  parser.hpp
-//  fomo_preproc
+//  Simdex
 //
 //  Created by Geet Sethi on 10/24/16.
 //  Copyright © 2016 Geet Sethi. All rights reserved.
@@ -11,10 +11,9 @@
 
 #include "types.hpp"
 #include <string>
-#include <stdio.h>
 
-float *parse_weights_csv(const std::string filename, const size_t num_rows,
-                         const size_t num_cols);
-int *parse_ids_csv(const std::string filename, const size_t num_rows);
+float *parse_weights_csv(const std::string filename, const int num_rows,
+                         const int num_cols);
+int *parse_ids_csv(const std::string filename, const int num_rows);
 
 #endif /* parser_hpp */
